Fix up the next auth bug
