+++ src/pages/users/index.tsx
@@ -1,7 +1,4 @@
 import Layout from '~/components/common/Layout'
-import { useSession } from 'next-auth/react'
-import LoadingScreen from '~/components/common/LoadingScreen'
-import { useRouter } from 'next/router'
 import Head from 'next/head'
 import DataTable from '~/components/users/Table'
 
@@ -16,17 +13,6 @@ const SEO = () => {
 }
 
 export default function AdminUserPage() {
-  const { data: session, status } = useSession({ required: true })
-  const router = useRouter()
-
-  if (status === 'loading') {
-    return <LoadingScreen />
-  }
-
-  if (!session.isAdmin) {
-    router.push('/users')
-  }
-
   return (
     <>
       <SEO />
